filtre et trie

+++ src/Controllers/Shop.php
@@ -10,24 +10,73 @@ class Shop extends Controller
     public function getAllShops()
     {
         $db = DB::getInstance()->getConnection();
-        $query = $db->prepare("SELECT * FROM shop");
-        $query->execute();
+        $data = json_decode(file_get_contents('php://input'), true);
+
+        $updatableFields = [];
+        $params = [];
+        if (!empty($data['name'])) {
+            $updatableFields[] = "name like :name";
+            $params['name'] = "%{$data['name']}%";
+        }
+        if (!empty($data['address'])) {
+            $updatableFields[] = "address like :address";
+            $params['address'] = "%{$data['address']}%";
+        }
+        if (!empty($data['zip']) && preg_match("/^[0-9]{5}$/", $data['zip'])) {
+            $updatableFields[] = "zip = :zip";
+            $params['zip'] = $data['zip'];
+        }
+        if (!empty($data['city'])) {
+            $updatableFields[] = "city like :city";
+            $params['city'] = "%{$data['city']}%";
+        }
+        $queryString = "SELECT * FROM shop ";
+        if (!empty($updatableFields)) {
+            $queryString .= "WHERE ";
+        }
+        $queryString .= implode(' AND ', $updatableFields);
+
+        if (isset($data['order'])) {
+            if (empty($data['order']['field']) && empty($data['order']['direction'])) {
+                $this->jsonResponse(["error" => "You must set field and direction parameters"], 400);
+                return;
+            }
+
+            if (!in_array($data['order']['field'], ['id', 'name', 'address', 'zip', 'city'])) {
+                $this->jsonResponse(["error" => "Invalid field parameter"], 400);
+                return;
+            }
+
+            if (!in_array($data['order']['direction'], ['ASC', 'DESC', 'asc', 'desc'])) {
+                $this->jsonResponse(["error" => "Invalid direction parameter, must be ASC or DESC"], 400);
+                return;
+            }
+            $queryString .= " ORDER BY {$data['order']['field']} {$data['order']['direction']}";
+        }
+
+        $query = $db->prepare($queryString);
+        $query->execute($params);
         $shop = $query->fetchAll(\PDO::FETCH_ASSOC);
 
         $this->jsonResponse($shop);
     }
 
-    public function getShop($id)
+    public function getShop(int $id): void
     {
         $db = DB::getInstance()->getConnection();
         $query = $db->prepare("SELECT * FROM shop WHERE id = :id");
         $query->execute(['id' => $id]);
         $shop = $query->fetchAll(\PDO::FETCH_ASSOC);
 
+        if (empty($shop)) {
+            $this->jsonResponse(["error" => "Shop with id $id not found"], 404);
+            return;
+        }
+
         $this->jsonResponse($shop);
     }
 
-    public function postShop()
+    public function postShop(): void
     {
         $data = json_decode(file_get_contents('php://input'), true);
 
@@ -58,10 +107,15 @@ class Shop extends Controller
         $this->jsonResponse($shop);
     }
 
-    public function putShop(int $id)
+    public function putShop(int $id): void
     {
         $data = json_decode(file_get_contents('php://input'), true);
         $db = DB::getInstance()->getConnection();
+
+        if (!$this->isShopExists($id)) {
+            $this->jsonResponse(["error" => "Shop with id $id not found"], 404);
+            return;
+        }
 
         $queryString = "UPDATE shop SET ";
 
@@ -93,15 +147,30 @@ class Shop extends Controller
         $query = $db->prepare("SELECT * FROM shop WHERE id = :id");
         $query->execute(['id' => $id]);
         $shop = $query->fetchAll(\PDO::FETCH_ASSOC);
+
         $this->jsonResponse($shop);
     }
 
-    public function deleteShop(int $id)
+    public function deleteShop(int $id): void
     {
         $db = DB::getInstance()->getConnection();
+
+        if (!$this->isShopExists($id)) {
+            $this->jsonResponse(["error" => "Shop with id $id not found"], 404);
+            return;
+        }
+
         $query = $db->prepare("DELETE FROM shop WHERE id = :id");
         $query->execute(['id' => $id]);
 
         $this->jsonResponse(["Shop $id deleted successfully!"]);
+    }
+
+    private function isShopExists(int $id): bool
+    {
+        $db = DB::getInstance()->getConnection();
+        $query = $db->prepare("SELECT COUNT(*) FROM shop WHERE id = :id");
+        $query->execute(['id' => $id]);
+        return (bool) $query->fetchColumn();
     }
 }
